requirements.txt
python==3.8
torch==2.0.0
torchvision==0.15.1
pillow==9.4.0
numpy==1.24.2
nltk==3.9.1
tqdm==4.66.5
transformers==4.44.0
evaluate==0.4.2